requirements.txt
matplotlib==3.5.1
numpy==1.19.2
Pillow==9.2.0
scikit_learn==1.1.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
